pyproject.toml
[project]
name = "lionzhd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aria2p>=0.12.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "inquirer>=3.4.0",
    "meilisearch>=0.31.5",
    "requests>=2.32.3",
    "uvicorn>=0.32.0",
    "whitenoise>=6.7.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[project.scripts]
lionz = "lionz:main"
web = "manage:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
version = { source = "scm" }

[tool.pdm.build]
package-dir = "."
source-includes = []
